# backend/VacciNationAPI/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "VacciNation Project Swagger Documentation",
    "version": "1.0.0",
    "title": "VacciNation"
  },
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/Appointment/all": {
      "get": {
        "tags": [
          "Appointments & Timeslots"
        ],
        "summary": "get all appointments",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "description": "baerer token for security and autorization",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "success - all appointments retrieved"
          },
          "400": {
            "description": "failed - some error was thrown"
          },
          "401": {
            "description": "failed - unauthorized"
          }
        }
      }
    },
    "/Appointment/all/open": {
      "get": {
        "tags": [
          "Appointments & Timeslots"
        ],
        "summary": "get all open appointments",
        "produces": [
          "application/json"
        ],
        "responses": {
          "202": {
            "description": "success - all open appointments retrieved"
          },
          "400": {
            "description": "failed - some error was thrown"
          }
        }
      }
    },
    "/Appointment/all/open/{vaccine_supplier}": {
      "get": {
        "tags": [
          "Appointments & Timeslots"
        ],
        "summary": "gets appointments of certain type",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "category_id",
            "description": "the timeslot filtered by categoryID (first dose, second dose, etc)",
            "type": "integer",
            "required": true
          },
          {
            "in": "path",
            "name": "vaccine_supplier",
            "description": "the timeslot filtered by supplier of the vaccine (Pfizer, Moderna, etc)",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "success - retrieves all open appointments from supplier retrieved"
          },
          "400": {
            "description": "failed - some error was thrown"
          }
        }
      }
    },
    "/Appointment/all/location/open": {
      "get": {
        "tags": [
          "Appointments & Timeslots"
        ],
        "summary": "gets open appointments at a certain location",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "location_id",
            "description": "the timeslot filtered by locationID",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "success - retrieves all open appointments at a location"
          },
          "400": {
            "description": "failed - some error was thrown"
          }
        }
      }
    },
    "/Appointment/all/location": {
      "get": {
        "tags": [
          "Appointments & Timeslots"
        ],
        "summary": "gets appointments at a certain location",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "description": "baerer token for security and autorization",
            "type": "string",
            "required": true
          },
          {
            "in": "query",
            "name": "location_id",
            "description": "the timeslot filtered by locationID",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "success - retrieves all appointments at a location"
          },
          "400": {
            "description": "failed - some error was thrown"
          },
          "401": {
            "description": "failed - unauthorized"
          }
        }
      }
    },
    "/Appointment/all/location/{supplier}": {
      "get": {
        "tags": [
          "Appointments & Timeslots"
        ],
        "summary": "gets appointments at a certain location with a certain vaccine type",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "description": "baerer token for security and autorization",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "supplier",
            "description": "the supplier of the vaccine",
            "type": "string",
            "required": true
          },
          {
            "in": "query",
            "name": "category_id",
            "description": "the timeslot filtered by categoryID",
            "type": "integer",
            "required": true
          },
          {
            "in": "query",
            "name": "location_id",
            "description": "the timeslot filtered by locationID",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "success - retrieves all appointments at a location in a category"
          },
          "400": {
            "description": "failed - some error was thrown"
          },
          "401": {
            "description": "failed - unauthorized"
          }
        }
      }
    },
    "/Appointment/all/location/open/{supplier}": {
      "get": {
        "tags": [
          "Appointments & Timeslots"
        ],
        "summary": "gets open appointments at a certain location with a certain vaccine type",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "description": "baerer token for security and autorization",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "supplier",
            "description": "the supplier of the vaccine",
            "type": "string",
            "required": true
          },
          {
            "in": "query",
            "name": "category_id",
            "description": "the timeslot filtered by categoryID",
            "type": "integer",
            "required": true
          },
          {
            "in": "query",
            "name": "location_id",
            "description": "the timeslot filtered by locationID",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "success - retrieves all open appointments at a location with a category/supplier"
          },
          "400": {
            "description": "failed - some error was thrown"
          },
          "401": {
            "description": "failed - unauthorized"
          }
        }
      }
    },
    "/Appointment/citizen/{id}": {
      "get": {
        "tags": [
          "Appointments & Timeslots"
        ],
        "summary": "get all appointments that a certain citizen has",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "description": "baerer token for security and autorization",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "id",
            "description": "the id of the citizen appointments will be filtered by",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "success - retrieves all appointments for citizen"
          },
          "400": {
            "description": "failed - some error was thrown"
          },
          "401": {
            "description": "failed - unauthorized"
          },
          "404": {
            "description": "failed - citizen was not found"
          }
        }
      }
    },
    "/Appointment/date": {
      "get": {
        "tags": [
          "Appointments & Timeslots"
        ],
        "summary": "get all appointments for a certain date (usecase example: staff get appointments for current day)",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "description": "baerer token for security and autorization",
            "type": "string",
            "required": true
          },
          {
            "in": "query",
            "name": "date",
            "description": "date value to search on - needs to be in format YYYY-MM-DD",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "success - retrieves all appointments for a date"
          },
          "400": {
            "description": "failed - some error was thrown"
          },
          "401": {
            "description": "failed - unauthorized"
          }
        }
      }
    },
    "/Appointment/open/date": {
      "get": {
        "tags": [
          "Appointments & Timeslots"
        ],
        "summary": "get all appointments for a certain date (usecase example: citizen get appointments for current day)",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "description": "date value to search on - needs to be in format YYYY-MM-DD",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "success - retrieves all open appointments for a date"
          },
          "400": {
            "description": "failed - some error was thrown"
          }
        }
      }
    },
    "/Appointment/date/{location_id}": {
      "get": {
        "tags": [
          "Appointments & Timeslots"
        ],
        "summary": "get all appointments for a certain date (usecase example: staff get appointments for current day) at a location",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "description": "baerer token for security and autorization",
            "type": "string",
            "required": true
          },
          {
            "in": "query",
            "name": "date",
            "description": "date value to search on - needs to be in format YYYY-MM-DD",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "location_id",
            "description": "location searching on",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "success - retrieves all appointments for a certain date at a location"
          },
          "400": {
            "description": "failed - some error was thrown"
          },
          "401": {
            "description": "failed - unauthorized"
          }
        }
      }
    },
    "/Appointment/date/open/{location_id}": {
      "get": {
        "tags": [
          "Appointments & Timeslots"
        ],
        "summary": "get all open appointments for a certain date (usecase example: staff get appointments for current day) at a location",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "description": "date value to search on - needs to be in format YYYY-MM-DD",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "location_id",
            "description": "location searching on",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "success - retrieves all open appointments for a certain date at a location"
          },
          "400": {
            "description": "failed - some error was thrown"
          }
        }
      }
    },
    "/Appointment/{id}": {
      "get": {
        "tags": [
          "Appointments & Timeslots"
        ],
        "summary": "get appointment by ID",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "description": "baerer token for security and autorization",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "id",
            "description": "get timeslot by this id",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "success - retrieves appointment with ID"
          },
          "400": {
            "description": "failed - some error was thrown"
          },
          "401": {
            "description": "failed - unauthorized"
          },
          "404": {
            "description": "failed - appointment not found"
          }
        }
      }
    },
...
  },
...
}